+++ R/priors.R
@@ -470,7 +470,8 @@ prior_string <- function(prior, ...) {
 #'
 #' @export
 get_prior <- function(formula, data, family = gaussian(), autocor = NULL,
-                      data2 = NULL, knots = NULL, sparse = NULL, ...) {
+                      data2 = NULL, knots = NULL, drop_unused_levels = TRUE,
+                      sparse = NULL, ...) {
   if (is.brmsfit(formula)) {
     stop2("Use 'prior_summary' to extract priors from 'brmsfit' objects.")
   }
@@ -485,7 +486,8 @@ get_prior <- function(formula, data, family = gaussian(), autocor = NULL,
   )
   data <- validate_data(
     data, bterms = bterms,
-    data2 = data2, knots = knots
+    data2 = data2, knots = knots,
+    drop_unused_levels = drop_unused_levels
   )
   .get_prior(bterms, data, ...)
 }
@@ -1159,13 +1161,14 @@ def_scale_prior.brmsterms <- function(x, data, center = TRUE, df = 3,
 #' @export
 validate_prior <- function(prior, formula, data, family = gaussian(),
                            sample_prior = "no", data2 = NULL, knots = NULL,
-                           ...) {
+                           drop_unused_levels = TRUE, ...) {
   formula <- validate_formula(formula, data = data, family = family)
   bterms <- brmsterms(formula)
   data2 <- validate_data2(data2, bterms = bterms)
   data <- validate_data(
     data, bterms = bterms,
-    data2 = data2, knots = knots
+    data2 = data2, knots = knots,
+    drop_unused_levels = drop_unused_levels
   )
   .validate_prior(
     prior, bterms = bterms, data = data,